version 1.1.1
 - Fix config settings inputs when it does not exist yet
 - Fix translations crash with some low resolutions
 - Fix chords note duration in translations
 - Copy star powers in translations
 - Fix inplace modifications when a file is used in clone hero or somewhere else
 - Skip translations when there's no expert track instead of crash

+++ src/splicer.ts
@@ -261,6 +261,8 @@ export class Splicer
                 let existingFiles = await fse.readdir(finalOutput);
                 for (let file of existingFiles)
                 {
+                    if (await fse.pathExists(this.output + "/" + file))
+                        continue;
                     try
                     {
                         await fse.remove(finalOutput + "/" + file);
@@ -271,7 +273,11 @@ export class Splicer
                     }
                 }
             }
-            await fse.copy(this.output, finalOutput, {overwrite: true})
+            let newFiles = await fse.readdir(this.output);
+            for (let file of newFiles)
+            {
+                await fse.promises.copyFile(this.output + "/" + file, finalOutput + "/" + file);
+            }
             this.output = finalOutput;
         }
     }
